[queue times] Actually use correct env vars (#5418)
* Follow up to https://github.com/pytorch/test-infra/pull/5417

I think the env var I was missing is the session token, which is needed
for temporary credentials. Instead of specifying that, I choose to not
explicitly set the credentials information and the `new S3Client` will
know to look for the env vars it needs on its own. This makes it
slightly easier to run locally if you have tokens already

A succeeding run is
https://github.com/pytorch/test-infra/actions/runs/9877586518/job/27279561554?pr=5418.
I got this by temporarily adding my branch to the iam trust relationship
policy and setting the workflow to run on push trigger.

I forgot that I could just add my branch there and test on a PR instead
of needing to test on the main branch

I don't think the upgrade to v3 is necessary but I decided to bundle
that in anyways