[hud] Add llm logic to log classifier
Summary:

This pr introduces the logic needed to apply llms to the log classifier.

Specifically the task that we're trying to do is that most of the time we hit the failure case with our log classifier, the actual error is somewhere above that error + nearby. The initial strategy for using llms is as follows.

1) Run the log classifier as normal
2) If we hit the failure case get a snippet of the log which contains n lines above the failure case.
3) Use the llm to find the failure line in the snippet.
4) Check if the failure line is legit otherwise use a better llm
5) If the better llm fails just give up and return the failure case.

This pr adds the logic such that we are able to perform steps 2-4. We still need to integrate this into the log classifier properly.

Current issues:
1) I'm running into rate limit issues while testing, so I assume I'll run into them in prod as well.
2) I think rust skills are rusty :P So error propogation is likely not done correctly here. Please comment in areas where this could be better.

Test Plan:

When uncommenting `test_make_query`, I get the following reasonable output

```
The query result is | mv: cannot stat 'td_results.json': No such file or directory
```