[log classifier] Add functionality to find multiple snippets
This pr adds functionality to get multiple snippets from a log.

Generally this is useful for when we want to feed the llm multiple snippets with errors in order to get an actionable error message.

In order to plug this into our current log classifier, I also added the `evaluate_ruleset_all_lines` function which gets all error line matches. The caveat of the function is that if a rule matches twice it will only return the bottomost match. However, I think this is fine for now as we want the llm to parse a variety of errors to figure out which is most important.