Allow full control on `python -m build` command (#5538)
This PR allows users to override the `python -m build` command when
building linux wheels.

I need this for torchcodec for 2 reasons.

1.  We want the build command to be:

```
python -m build
```

instead of the currently hardcoded:

```
python -m build --wheel
```

From the `build` docs, the default behavior is to *not* pass `--wheel`,
so that the wheel is built *from* the sdist. This is slightly safer as
it ensures that a sdist can indeed be built.


2. We want to pass an environment variable, i.e. we want to write


```
SOME_ENV_VAR=123 python -m build ...
```

I realize this can be done with `env-var-script`. But using
`env_var_script` leads to a decentralized build command, and it's more
difficult to really understand what command is being used for building.
It would be simpler and easier to grasp if we could write the whole
command as a one-liner like above.