tools: Add script to run clickhouse formatter
Adds a script to run the clickhouse formatter on all of our clickhouse
queries.

Changes should be automatically made to the respective `.sql` files
within the torchci/clickhouse_queries directory and we can utilize git
to track what the changes are.

NOTE: This does not install the clickhouse binary for you and will need
to install it on their own.

Usage:

> bash .github/scripts/run_clickhouse_format.sh

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>