[disablebot] Exclusion for oncall: pt2 as removable platform label (#5578)
The disable bot automatically tries to delete labels it thinks is no
longer relevant. Usually it makes sense, like removing module: windows
if the platforms change to not include windows. Unfortunately it also
puts oncall: pt2 in this category, but oncall: pt2 can also be a test
owner label, so it should stay