Allow setting minimum available runners to 0
This change modifies the check to allow the minimum available runners
setting to be 0 while keeping the default value to 10. The sanity check to
ensure that at least a positive number is used remains. If an invalid
number is passed then the setting is assumed to be 0.

Scale-up function however requires a minimum of at least 1 to be able
to do its calculations on how many runners to scale up. A value of 0
here breaks the logic and causes the function to never create any runners.
So logic is added here to ensure that scale-up's min value is 1.

Issue: https://lf-pytorch.atlassian.net/browse/PC-33
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>