Set the AWS credential duration in mobile job (#5685) (#5702)
As reported by @guangy10 in

https://github.com/pytorch/executorch/pull/5441#issuecomment-2357004155,
it turns out that the job timeout can be set to a higher value (120
minutes), but the AWS credential still expires after 1 hour
https://github.com/aws-actions/configure-aws-credentials.

The fix here is to set the role duration correctly. However, the max
that it can go is 18000 seconds. I believe that's plenty of room
already, but if a higher is needed, we will need to update the server
side

https://github.com/pytorch-labs/pytorch-gha-infra/blob/main/runners/gha_roles.tf#L1425

### Testing

The role duration is set correctly to 7200 seconds (2 hours)
https://github.com/pytorch/test-infra/actions/runs/10916800466/job/30298898748#step:4:4