[drci] Update comments that say pending (#5699)
Dr. CI updates comments every 15 minutes (done via GHA) and queries for
the last 30 minutes of completed jobs in order to account for small
errors like Dr. CI not actually running due to GHA not scheduling in
time.

However, if Dr. CI fails to run multiple times in a row or webhooks are
sent late (ex there was an outage), we can get cases where the job
completes but Dr. CI never picks it up, so the comment always says that
there is something pending.

One way to solve this (implemented here) is to also query for comments
that say there are jobs pending and update those too.

Currently, it searches for any comments that say there are things
pending in the last month (do we want to make this interval smaller?)
and Dr. CI will update those as well.

Drawbacks:
* additional CH queries
* probably a lot of overlap between jobs that finished in the last 30
minutes and comments that say pending