Fix wrong commit date on benchmark dashboards (#5725)
Fixes https://github.com/pytorch/test-infra/issues/5696

I discover two issues with the current query to get the commits and
their date:

* The query inherits the granularity parameter, so when the granularity
is set to week, for example, all commit's dates are truncated to the
beginning of the week. This is the root cause of
https://github.com/pytorch/test-infra/issues/5696. For more context,
switching to weekly granularity is a hack I have at the moment to be
able to reduce the payload side and be able to load last quarter data
(https://github.com/pytorch/test-infra/issues/4123). So the issue only
manifests when a time range from last quarter or more is selected. The
fix then is just to truncate commit's dates by day because that's the
format displayed in the dashboard (YYYY/MM/DD)
* Truncating the commit's date by day also fixes an off-by-one-day
mismatch in the dashboard. For example,
[77fba0c407](https://hud.pytorch.org/pytorch/pytorch/commit/77fba0c407c10e3e2b5182578aa8affae6abc398#inductor-A100-perf-nightly)'s
date is 2024/09/26, which is actually a day later than the actual date
https://github.com/pytorch/pytorch/commit/77fba0c407c10e3e2b5182578aa8affae6abc398
(2024/09/25)

### Testing

* Selecting last quarter with weekly granularity doesn't screw up the
commit's dates now
https://torchci-git-fork-huydhn-fix-wrong-commit-da-ccea24-fbopensource.vercel.app/benchmark/compilers?dashboard=torchinductor&startTime=Sat%2C%2029%20Jun%202024%2005%3A17%3A28%20GMT&stopTime=Fri%2C%2027%20Sep%202024%2005%3A17%3A28%20GMT&granularity=week&suite=torchbench&mode=inference&dtype=bfloat16&deviceName=cuda%20(a100)&lBranch=main&lCommit=77fba0c407c10e3e2b5182578aa8affae6abc398&rBranch=main&rCommit=a676b7c5f3cf05690afbce3c99575916b3b468aa
* Regular view
https://torchci-git-fork-huydhn-fix-wrong-commit-da-ccea24-fbopensource.vercel.app/benchmark/compilers
shows the correct commit's dates.
* Other benchmark dashboards also looks correct:
*
https://torchci-git-fork-huydhn-fix-wrong-commit-da-ccea24-fbopensource.vercel.app/benchmark/llms