Migrate get-workflow-job-id GHA to test-infra (#5740)
Copied from pytorch. This is needed to get the job ID on ExecuTorch
mobile benchmark job. The problem I'm trying to solve is to find the
artifacts from parallel jobs with matrix strategy. Atm, GitHub doesn't
offer a way to get this information
https://github.com/orgs/community/discussions/17245. One workaround
solution is to save the artifact with an unique name using the job ID,
then try to download that artifact later.

I will switch PT to use this in another PR.