Excpect scale-configs to all live in test-infra now (#5788)
Part of the workflow to move the LF scale-config.yml files from
pytorch/pytorch to test-infra (details in
https://github.com/pytorch/test-infra/pull/5767)

This updates the validation script to no longer expect to update the
pytorch/pytorch version of these scale configs. It resulted in many
other aspects of that file becoming simpler as well.