[CH] Migrate failure annotation page (#5803)
This is more like a refactoring, not a migration, but it's needed to
make the query works.

* `failed_workflow_jobs` runs OOM on ClickHouse trying to load all
failed jobs from commits that are having more than N failures. I rewrite
the logic to get rid of N because we only set it to 0 anyway. The
simplified logic can be then be run
* https://hud.pytorch.org/failedjobs/pytorch/pytorch/main has duplicated
entries. This is a subtle bug from
`torchci/components/JobAnnotationToggle.tsx` that took me way too long
to figure out.
* There are several React warnings on the page, so I fix them too

### Testing


https://torchci-git-fork-huydhn-ch-migrate-job-anno-45acb4-fbopensource.vercel.app/failedjobs/pytorch/pytorch/main