[HUD] Fix workflow dispatcher (#5886)
Dunno when this broke, it was before the next update, but fix workflow
dispatcher on HUD

`return void stuff` because after the next update, I was getting `API
handler should not return a value`. Next wanted it to return undefined,
and void somethng is the same as evaluate something and then return
undefined. Ref: https://github.com/vercel/next.js/discussions/48951