[HUD] Button for updating Dr. CI (#5844)
Add button on top right where logged in users can update Dr. CI
https://torchci-git-csl-drcibutton-fbopensource.vercel.app/pr/138513

I am failing to log in on the preview, but it works locally, probably
due to prod vs preview env vars

Rate limit users to 10 calls / hr

Is this a good idea? No clue

Should this be limited to people with write access?