Login to docker.io (#5957)
This is an attempt to mitigate the rate limit issue we are seeing when
pulling the base image from docker.io
https://docs.docker.com/docker-hub/download-rate-limit/#rate-limit.

Here is the summary of my support ticket to GitHub about the issue:

1. This behavior is from Docker Hub itself and isn't something GitHub
can directly control. You might consider a paid subscription to Docker
Hub which increases the rate limit. You would need to pre-authenticate
to Docker Hub on these runners before the job starts as GitHub builds
the Docker containers before the first explicit step on the job.
2. Alternatively, we could move the images to https://ghcr.io which does
not have such rate limits.
3. GitHub has an agreement with Docker so that the rate limits do not
apply for publicly accessible images downloaded through a GitHub-hosted
runner but this does not apply to self-hosted runners.
 
This is basically the first approach. We could try it out and see if it
helps. The worst thing that could happen is to be rate limited again.

This goes together with
https://github.com/pytorch-labs/pytorch-gha-infra/pull/544