Bugfix: Overprovision scaleups on ephemeral instances & extend runner scale up code to enable larger scale ups (#6015)
Part of https://github.com/pytorch/test-infra/issues/5798

### Changes included
- Fix a big bug with how we provisioned instances for ephemeral
instances. In the old code, the fewer ephemeral runners we had
available, the _less_ likely we were to overprovision them, which sounds
like the opposite of what we actually want to do. Changed that logic to
more explicitly give ephemeral runners a 50% chance of overprovisioning
the incoming request by two instances, which is what I think [this
PR](https://github.com/pytorch/test-infra/pull/5501) was intending to do
(we can tweak these numbers as desired). This fix might handle most of
our ephemeral runner woes by itself...
- Extend the allRunnersBusy function to enable it to handle scale up
requests of more than a single instance (which will be leveraged in a
future PR)
- Renames allRUnnersBusy to something more accurate name