[BE] [PyTorchBot] More understandable error  message when labels cannot be added (#6088)
Fixes https://github.com/pytorch/test-infra/issues/5900

When a `ciflow` label is added to a PR, we normally trigger workflows
against it. However, if the PR has not yet been authorized to actually
run workflows then we do not allow the label to be added. (This feature
was added by https://github.com/pytorch/test-infra/pull/4892)

This change improves the reasoning given by the error messages, making
it easier for the reviewer to understand why a label was not added to
the PR and the steps they need to take to actually add the desired
label.