Add speedup metric for TorchAO (#6118)
This is my initial attempt to add the speedup metric for TorchAO. This
is done by comparing the gain of `autoquant` v.s. `noquant`. This is by
no means the best approach because it requires custom logic for TorchAO
on the dashboard. On the other hand, it's easy to implement and I think
it's better to have the UX done first to gather early feedbacks from
@jerryzh168 and the rest of ao team first.

IMO, better approaches would be to either 1) set the speedup metric on
TorchAO side or 2) compute the speed up metric on ClickHouse. Both are
more involved and requires further design discussion.

### Testing


https://torchci-git-fork-huydhn-add-speedup-llm-dashboard-fbopensource.vercel.app/benchmark/llms?repoName=pytorch%2Fao