Add a chart to show the total the number of models (#6169)
Per the char with @zou3519 and @shunting314, I'm adding another chart to
show the total number of models covered in the benchmark run. This is
the denominator used to compute the pass rate. I'm showing this in a
separate chart because there is no easy way to include this information
in the current pass rate chart given the current implementation.

I also increase the legend padding a bit to make the chart easier to
view.

### Testing


https://torchci-git-fork-huydhn-show-inductor-passr-b83539-fbopensource.vercel.app/benchmark/compilers?dashboard=torchinductor&startTime=Thu%2C%2018%20Jul%202024%2022%3A52%3A52%20GMT&stopTime=Tue%2C%2014%20Jan%202025%2023%3A52%3A52%20GMT&granularity=week&mode=inference&dtype=bfloat16&deviceName=cuda%20(a100)&lBranch=main&lCommit=1dab79470dbecef79ba4c7d4308d8a181091e58e&rBranch=main&rCommit=b732b52f1e4378f8486ceb5e7026be3321c2651c