Use oss_ci_benchmark_metadata materialized view (#6167)
I'm attempting to use a new materialized view called
`oss_ci_benchmark_metadata` to make it faster to query benchmark
metadata. I have already added the new view manually but its definition
is included in this PR for review.

The old query works, but it's slower because the original table doesn't
keep several important columns, i.e. benchmark name, in order
https://github.com/pytorch/test-infra/blob/main/torchci/clickhouse_queries/oss_ci_benchmark_v3/query.sql#L98-L107.
According to
https://clickhouse.com/docs/en/sql-reference/statements/alter/order-by,
it seems that we cannot add existing columns into ORDER BY.

### Testing

When loading the benchmark metadata:

* Slooooow
https://hud.pytorch.org/benchmark/llms?repoName=pytorch%2Fpytorch
* Faster
https://torchci-git-fork-huydhn-use-osscibenchmarkm-96b1e1-fbopensource.vercel.app/benchmark/llms?repoName=pytorch%2Fpytorch

Loading the benchmark data is still relatively slow IMO, but I will try
to improve that in a separate PR.