Improves SLI page by adding sharable stateful link and sets 'pet, ephemeral, nonephemeral' as default options (#6202)
Adding a sharable link is a useful thing for when we want to share and
discuss things inside the team (easier than sharing print screens most
of the time).

All is not that useful of a metric for us, it include 3rd party provided
runners and github runners. This makes the information noisy and
non-actionable. To focus on what we manage and provide some useful
breakdown I opted to have the following breakdowns by default:

- "pet" - include all our pet instances, macos and the shared instances
with A100 and H100;
- "ephemeral" - are all instances that are dynamically scaled both by
Meta and LF, that are ephemeral. They are very fragile to stockouts;
- "nonephemeral" - are all instances that are dynamically scaled both by
Meta and LF, that are non-ephemeral. They are somewhat resilient to
stockouts;

Other breakdowns are more useful as troubleshoot. More details can be
found checking the code:
https://github.com/pytorch-labs/pytorch-gha-infra/pull/590/files#diff-9837b2a66265c20a3b89d979d8c446108a824c8af3f997f535d62f644089e82cR148