Local results and perf comparisons for queries (#6206)
Second commit onwards are the actual changes. The first commit just
changes all the formatting for the existing files.

This changes the formatting of the params.json file to allow for example
inputs for the queries that can be used to compare results and
performance to be used in the script mentioned below. The new format can
be seen in the README.

This also adds a script to get compare time and memory a query uses
between local and a specific commit. It can also compare what the query
returns.