Move meta canary scale-config.yml file to this repository (#6266)
## Added `.github/canary-scale-config.yml` file. 

This file should be the one used by Pytorch Canary on Meta's
environment.

Its been painful to develop/test things using pytorch/pytorch-canary on
meta's fleet. The main reason for this is the need to add/update the
scale-config.yml file every time a test needs to be executed. This
always generates merge conflicts with pytorch/pytorch and requires a
substantial manual work.

The other advantage of having this file here is the clarity of where all
the configuration is, so it is central in a single place for all 4
environments that we have.

Finally, the validation scripts should then ensure the quality of the
code and that all 4 scale-config files are in sync.

## Using the standard `c.<something>` 

The usage of the standard `c.<something>` in detriment of
`<something>.canary` for Meta's canary runners, is beneficial. Reduced
names sizes, a single standard for all our runners, and greater
compatibility with current implementations. This should also leverage
and empower in the future to streamline pytorch/pytorch-canary CI to run
with canary runners automatically, without the need for the boring
change of runners definitions in all workflows when opening the PR.