Filter out devices that are not run (#6277)
Fixes https://github.com/pytorch/executorch/issues/7986 (for real this
time)

The problem is that the records in the benchmark database alone don't
have the information to differentiate between benchmarks that are failed
to run and benchmarks that are not run. Both show up as 0 on the
dashboard. Note that we can do a join with `workflow_job` table to get
this information, but it's a rather slow and expensive route. So, I opt
for a quicker approach to keep track of the list of valid devices on the
dashboard side. A valid device is one that is run by the selected commit
and has at least one non-zero record there.

### Testing


https://torchci-git-fork-huydhn-better-model-filter-fbopensource.vercel.app/benchmark/llms?repoName=pytorch%2Fexecutorch