Fix summary table for same commit on different devices (#6343)
This is a bug reported by @malfet where the dashboard only shows the
value from the new commit (right) when the left commit and the right
commit are the same but on different devices. This defeats the purpose
of comparing different devices.

The bug here is that the `l → r` display worked under the assumption
that same commit mean same benchmark, which is not true anymore. Same
benchmark now means same commit plus same device.

### Preview


https://torchci-git-fork-huydhn-fix-cross-device-sa-e8d1e5-fbopensource.vercel.app/benchmark/compilers?dashboard=torchinductor&startTime=Wed%2C%2019%20Feb%202025%2021%3A06%3A10%20GMT&stopTime=Wed%2C%2026%20Feb%202025%2021%3A06%3A10%20GMT&granularity=hour&mode=inference&dtype=bfloat16&lDeviceName=rocm%20(mi300x)&rDeviceName=cuda%20(a100)&lBranch=main&lCommit=d0adff761ea4d299528c89bee23008dbc622846e&rBranch=main&rCommit=d0adff761ea4d299528c89bee23008dbc622846e