[bots] Checking for approved workflow runs: handle runs being startup failure (#6329)
If all the runs are startup failure, the function would return true.
However, this can hide the need for those runs to be approved. This
might result in runs being rejected when they shouldn't, but I doubt an
actual PR is going to mangle workflows and still want CI

Err on the side of caution by refusing if any job has a startup failure,
even if there are approved runs

Example: https://github.com/pytorch/pytorch/pull/147774

(I am hash update bot, its a poorly named account since I was originally
going to use for something else and was the fastest account I could get
right now to test this)

There is a startup_failure conclusion, but depending on how you mess up
the workflow syntax, you can either get failure or startup_failure
(https://github.com/pytorch/pytorch/pull/148016/files), and it seems
like if it is going to be a startup_failure if it runs, the workflow
doesn't show up at all when you query about the commit