[HUD] Restore tts page jobName router query (#6360)
The tts and job duration tables in the metrics page use jobName in the
link to quickly specify a job. I got rid of it in
https://github.com/pytorch/test-infra/pull/6326 in favor of a compressed
list so you could permalink multiple jobs and forgot that the metrics
page linked here. This PR restores the functionality and allows both
jobNames and jobNamesCompressed